Module Incr_dom

include module type of struct include Virtual_dom end

module Vdom = Virtual_dom.Vdom

module Dom_float = Virtual_dom.Dom_float

module Top_level_effects = Virtual_dom.Top_level_effects

module App_intf : sig ... end

module Incr : sig ... end

module Js_misc : sig ... end

This module has a small collection of helpful bits of javascript that have no other obvious home.

module Start_app : sig ... end

For starting up an Incr_dom app.

module Component : sig ... end

A Component captures the basic operations that need to be provided in order to build an Incr_dom app. The same type can be used both at the top-level of an app, as well as for defining individual components within a larger app.