Module Vdom_file_download

A utility for causing the browser to download a file with the contents taken from a string.

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val create : filename:string -> mimetype:string -> contents:string -> t

val trigger : t -> unit

Immediately cause the browse to trigger a download for t. You can use this to build a more complex UI for downloads than is offered by Button.create -- e.g. a button that calls an RPC and downloads a file based on the result.

In tests, this just prints a message rather than triggering the download.

module Button : sig ... end